# Use compiled regex for common pack delimiter parsing