#17594 

This change uses compiled regex when parsing common pack delimiters.  Here are the benchmark results.

Test:
```
func BenchmarkFindPackDelimiterStringCommon(b *testing.B) {
	// Input data for benchmarking
	input := "pack/Global/Foo"

	// Run the benchmark
	b.ResetTimer()
	for i := 0; i < b.N; i++ {
		findPackDelimiterString(input)
	}
}
```

Before:
```
BenchmarkFindPackDelimiterStringCommon-12         406230              3163 ns/op            5031 B/op         39 allocs/op
PASS
ok      github.com/fleetdm/fleet/v4/server/service      1.749s
```

After:
```
BenchmarkFindPackDelimiterStringCommon-12    	 1312630	       911.2 ns/op	     337 B/op	       4 allocs/op
PASS
ok  	github.com/fleetdm/fleet/v4/server/service	2.500s
```


- [X] Added/updated tests
- [X] Manual QA for all new/changed functionality